+++ iOS/AppboyReactBridge/AppboyReactBridge/AppboyReactBridge.m
@@ -456,6 +456,11 @@ RCT_EXPORT_METHOD(requestLocationInitialization) {
   RCTLogInfo(@"Warning: This is an Android only feature.");
 }
 
+RCT_EXPORT_METHOD(requestGeofences:(double)latitude longitude:(double)longitude) {
+  RCTLogInfo(@"[Appboy sharedInstance] requestGeofencesWithLongitude:latitude: with latitude %g and longitude %g", latitude, longitude);
+  [[Appboy sharedInstance] requestGeofencesWithLongitude:longitude latitude:latitude];
+}
+
 RCT_EXPORT_METHOD(getCardCountForCategories:(NSString *)category callback:(RCTResponseSenderBlock)callback) {
   ABKCardCategory cardCategory = [self getCardCategoryForString:category];
   if (cardCategory == 0) {